chore: Split `send` module
