+++ src/lib.rs
@@ -3,7 +3,8 @@
 pub mod client;
 mod handle;
 mod receive;
-mod send;
+mod send_command;
+mod send_response;
 pub mod server;
 pub mod stream;
 pub mod types;
